feat(newsletter): rename to NewsletterForm
- simplify to only have rendering and submitting form
- move thank you to CtaRegion
- add compact option for sidebar
- various theme and style changes
