+++ components/CtaRegion/components/CtaMessageNewsletter/CtaMessageNewsletter.tsx
@@ -3,21 +3,26 @@
  * Component for newsletter CTA messages.
  */
 
-import React from 'react';
+import React, { useState } from 'react';
 import {
+  Box,
+  ButtonProps,
   Card,
   CardHeader,
   CardContent,
   CardActions,
-  ButtonProps,
+  Divider,
+  Grid,
   Typography,
   ThemeProvider
 } from '@material-ui/core';
-import { Newsletter } from '@components/Newsletter';
+import { CheckCircleOutlineSharp } from '@material-ui/icons';
+import { NewsletterForm } from '@components/NewsletterForm';
 import { ICtaMessageProps } from '@interfaces/cta';
 import { ctaMessageNewsletterTheme } from './CtaMessageNewsletter.styles';
 
 export const CtaMessageNewsletter = ({ data }: ICtaMessageProps) => {
+  const [subscribed, setSubscribed] = useState(false);
   const {
     heading,
     message,
@@ -35,6 +40,10 @@ export const CtaMessageNewsletter = ({ data }: ICtaMessageProps) => {
     disableElevation: true
   };
 
+  const handleSubscribed = () => {
+    setSubscribed(true);
+  };
+
   return (
     <ThemeProvider theme={ctaMessageNewsletterTheme}>
       <Card elevation={0}>
@@ -50,13 +59,51 @@ export const CtaMessageNewsletter = ({ data }: ICtaMessageProps) => {
           </CardContent>
         )}
         {hasActions && (
-          <CardActions>
-            <Newsletter
-              label={action && action.name}
-              options={newsletterOptions}
-              buttonProps={actionAttrs}
-            />
-          </CardActions>
+          <>
+            <Divider />
+            <CardActions>
+              {!subscribed && (
+                <NewsletterForm
+                  label={action && action.name}
+                  options={newsletterOptions}
+                  onSubscribed={handleSubscribed}
+                />
+              )}
+              {subscribed && (
+                <Grid
+                  container
+                  spacing={4}
+                  justify="center"
+                  alignContent="center"
+                >
+                  <Grid item xs={12}>
+                    <Box
+                      display="grid"
+                      gridTemplateColumns="min-content 1fr"
+                      gridGap={16}
+                      justifyContent="center"
+                      alignItems="center"
+                    >
+                      <Box>
+                        <CheckCircleOutlineSharp
+                          color="primary"
+                          style={{ fontSize: '5rem' }}
+                        />
+                      </Box>
+                      <Box>
+                        <Typography variant="overline">
+                          You&apos;re Subscribed
+                        </Typography>
+                        <Typography>
+                          <strong>Thank you!</strong> You mean the world to us.
+                        </Typography>
+                      </Box>
+                    </Box>
+                  </Grid>
+                </Grid>
+              )}
+            </CardActions>
+          </>
         )}
       </Card>
     </ThemeProvider>
