+++ components/Story/layouts/index.ts
